+++ converter.h
@@ -1,8 +1,10 @@
 #ifndef CONVERTER_H
 #define CONVERTER_H
 
-int base_conversion (int input, int inputbase, int outputbase);
-int decimal2base (int input, double base);
-int base2decimal (int input, double base);
+#include <QtCore>
+
+qint64 base_conversion (qint64 input, qint64 inputbase, qint64 outputbase);
+qint64 decimal2base (qint64 input, double base);
+qint64 base2decimal (qint64 input, double base);
 
 #endif
